Phi_data/requirements.txt
phidata
openai
duckdb
duckduckgo-search
pydantic
pandas
wikipedia
sqlalchemy
pgvector
pypdf
psycopg
arxiv
yfinance
ollama
